requirements.txt
absl-py==0.9.0
alabaster==0.7.12
astor==0.8.1
astroid==2.4.1
atomicwrites==1.4.0
attrs==19.3.0
autopep8==1.5.2
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.5
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.4.1
colorama==0.4.3
coremltools==3.3
cryptography==2.9.2
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
docutils==0.16
entrypoints==0.3
flake8==3.7.9
gast==0.3.3
grpcio==1.28.1
h5py==2.10.0
helpdev==0.7.1
idna==2.9
imageio==2.8.0
imagesize==1.2.0
imgaug==0.4.0
importlib-metadata==1.6.0
intervaltree==3.0.2
ipykernel==5.2.1
ipyparallel==6.3.0
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.15.2
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.0.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==20.0.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.24.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==2.2.2
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nose==1.3.7
notebook==6.0.3
numba==0.39.0
numpy==1.18.1
numpydoc==0.9.2
opencv-python==4.2.0.34
packaging==20.3
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
pycocotools==2.0
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==5.0.2
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.7.2
PyQtWebEngine==5.12.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole==4.7.3
QtPy==1.9.0
requests==2.23.0
rope==0.17.0
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.3
spyder-kernels==1.9.1
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
yapf==0.30.0
zipp==1.2.0
